requirements.txt
aniso8601==1.3.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
python-dateutil==2.6.1
python-zillow==0.1.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
xmltodict==0.11.0
